Move cut and copy out of MainTable into LibraryTab
