+++ src/main/java/org/jabref/gui/LibraryTab.java
@@ -1,5 +1,6 @@
 package org.jabref.gui;
 
+import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Collections;
 import java.util.List;
@@ -63,7 +64,6 @@ import org.jabref.gui.util.BackgroundTask;
 import org.jabref.gui.util.OptionalObjectProperty;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.gui.util.UiTaskExecutor;
-import org.jabref.logic.ai.AiService;
 import org.jabref.logic.citationstyle.CitationStyleCache;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
@@ -82,6 +82,7 @@ import org.jabref.model.database.event.EntriesRemovedEvent;
 import org.jabref.model.entry.Author;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibEntryTypesManager;
+import org.jabref.model.entry.BibtexString;
 import org.jabref.model.entry.LinkedFile;
 import org.jabref.model.entry.event.EntriesEventSource;
 import org.jabref.model.entry.event.FieldChangedEvent;
@@ -118,7 +119,6 @@ public class LibraryTab extends Tab {
     private final CountingUndoManager undoManager;
     private final DialogService dialogService;
     private final PreferencesService preferencesService;
-    private final AiService aiService;
     private final FileUpdateMonitor fileUpdateMonitor;
     private final StateManager stateManager;
     private final BibEntryTypesManager entryTypesManager;
@@ -174,7 +174,6 @@ public class LibraryTab extends Tab {
                        LibraryTabContainer tabContainer,
                        DialogService dialogService,
                        PreferencesService preferencesService,
-                       AiService aiService,
                        StateManager stateManager,
                        FileUpdateMonitor fileUpdateMonitor,
                        BibEntryTypesManager entryTypesManager,
@@ -187,7 +186,6 @@ public class LibraryTab extends Tab {
         this.undoManager = undoManager;
         this.dialogService = dialogService;
         this.preferencesService = Objects.requireNonNull(preferencesService);
-        this.aiService = Objects.requireNonNull(aiService);
         this.stateManager = Objects.requireNonNull(stateManager);
         this.fileUpdateMonitor = fileUpdateMonitor;
         this.entryTypesManager = entryTypesManager;
@@ -948,7 +946,26 @@ public class LibraryTab extends Tab {
     }
 
     public void copy() {
-        mainTable.copy();
+        List<BibEntry> selectedEntries = getSelectedEntries();
+
+        if (!selectedEntries.isEmpty()) {
+            List<BibtexString> stringConstants = getUsedStringValues(selectedEntries);
+            try {
+                if (stringConstants.isEmpty()) {
+                    clipBoardManager.setContent(selectedEntries, entryTypesManager);
+                } else {
+                    clipBoardManager.setContent(selectedEntries, entryTypesManager, stringConstants);
+                }
+                dialogService.notify(Localization.lang("Copied %0 entry(ies)", selectedEntries.size()));
+            } catch (
+                    IOException e) {
+                LOGGER.error("Error while copying selected entries to clipboard.", e);
+            }
+        }
+    }
+
+    private List<BibtexString> getUsedStringValues(List<BibEntry> entries) {
+        return bibDatabaseContext.getDatabase().getUsedStrings(entries);
     }
 
     public void paste() {
@@ -960,7 +977,8 @@ public class LibraryTab extends Tab {
     }
 
     public void cut() {
-        mainTable.cut();
+        copy();
+        delete(StandardActions.CUT);
     }
 
     public boolean isModified() {
@@ -991,7 +1009,6 @@ public class LibraryTab extends Tab {
                                               Path file,
                                               DialogService dialogService,
                                               PreferencesService preferencesService,
-                                              AiService aiService,
                                               StateManager stateManager,
                                               LibraryTabContainer tabContainer,
                                               FileUpdateMonitor fileUpdateMonitor,
@@ -1007,7 +1024,6 @@ public class LibraryTab extends Tab {
                 tabContainer,
                 dialogService,
                 preferencesService,
-                aiService,
                 stateManager,
                 fileUpdateMonitor,
                 entryTypesManager,
@@ -1029,7 +1045,6 @@ public class LibraryTab extends Tab {
                                               LibraryTabContainer tabContainer,
                                               DialogService dialogService,
                                               PreferencesService preferencesService,
-                                              AiService aiService,
                                               StateManager stateManager,
                                               FileUpdateMonitor fileUpdateMonitor,
                                               BibEntryTypesManager entryTypesManager,
@@ -1043,7 +1058,6 @@ public class LibraryTab extends Tab {
                 tabContainer,
                 dialogService,
                 preferencesService,
-                aiService,
                 stateManager,
                 fileUpdateMonitor,
                 entryTypesManager,
